requirements.txt
Pillow==10.0.0
ttkthemes==3.2.2
requests==2.31.0
python-dotenv==1.0.0
ollama==0.4.8
